+++ src/App/Component/Vote/Exceptions/UpvoteFailedException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Component\Vote\Exceptions;
+
+/**
+ * The exception thrown when trying to upvote an invalid world record
+ */
+class UpvoteFailedException extends \Exception
+{
+}